Rename and move files to avoid name clashes
