+++ src/OpenLoco/src/CompanyManager.h
@@ -2,7 +2,7 @@
 
 #include "Company.h"
 #include "Core/LocoFixedVector.hpp"
-#include "Limits.h"
+#include "Engine/Limits.h"
 #include "Map/Map.hpp"
 #include "Types.hpp"
 #include <array>
