+++ src/OpenLoco/src/Engine/Limits.h
