+++ src/OpenLoco/src/GameCommands/RenameCompanyOwner.cpp
@@ -1,9 +1,9 @@
 #include "CompanyManager.h"
 #include "Economy/Expenditures.h"
+#include "Engine/Limits.h"
 #include "GameCommands.h"
 #include "Graphics/Gfx.h"
 #include "Interop/Interop.hpp"
-#include "Limits.h"
 #include "Localisation/FormatArguments.hpp"
 #include "Localisation/StringIds.h"
 #include "Localisation/StringManager.h"
