+++ src/OpenLoco/src/Industry.h
@@ -2,7 +2,7 @@
 
 #include "Core/BitSet.hpp"
 #include "Core/Span.hpp"
-#include "Limits.h"
+#include "Engine/Limits.h"
 #include "Map/Tile.h"
 #include "Map/TileLoop.hpp"
 #include "Types.hpp"
