+++ src/OpenLoco/src/IndustryManager.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "Core/LocoFixedVector.hpp"
+#include "Engine/Limits.h"
 #include "Industry.h"
-#include "Limits.h"
 #include <array>
 #include <cstddef>
