+++ src/OpenLoco/src/Map/WaveManager.cpp
@@ -1,9 +1,9 @@
 #include "WaveManager.h"
 #include "Core/LocoFixedVector.hpp"
+#include "Engine/Limits.h"
 #include "Game.h"
 #include "GameState.h"
 #include "Interop/Interop.hpp"
-#include "Limits.h"
 #include "ScenarioManager.h"
 #include "SurfaceElement.h"
 #include "TileManager.h"
