+++ src/OpenLoco/src/Objects/ObjectManager.h
@@ -2,7 +2,7 @@
 
 #include "Core/Span.hpp"
 #include "Object.h"
-#include "Ui/Types.hpp"
+#include "Ui/UiTypes.hpp"
 #include <optional>
 #include <vector>
