+++ src/OpenLoco/src/S5/S5.h
@@ -2,7 +2,7 @@
 
 #include "Core/FileSystem.hpp"
 #include "EditorController.h"
-#include "Limits.h"
+#include "Engine/Limits.h"
 #include "Objects/Object.h"
 #include "ScenarioObjective.h"
 #include <cstdint>
