+++ src/OpenLoco/src/StationManager.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "Core/LocoFixedVector.hpp"
-#include "Limits.h"
+#include "Engine/Limits.h"
 #include "Station.h"
 #include <array>
 #include <cstddef>
