+++ src/OpenLoco/src/TownManager.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "Core/LocoFixedVector.hpp"
-#include "Limits.h"
+#include "Engine/Limits.h"
 #include "Town.h"
 #include <array>
 #include <optional>
