+++ src/OpenLoco/src/Ui/Dropdown.cpp
@@ -1,10 +1,10 @@
 #include "Dropdown.h"
 #include "CompanyManager.h"
 #include "Console.h"
+#include "Engine/Limits.h"
 #include "Graphics/ImageIds.h"
 #include "Input.h"
 #include "Interop/Interop.hpp"
-#include "Limits.h"
 #include "Localisation/FormatArguments.hpp"
 #include "Localisation/StringIds.h"
 #include "Objects/CompetitorObject.h"
