+++ src/OpenLoco/src/Ui/Rect.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Types.hpp"
+#include "UiTypes.hpp"
 #include <algorithm>
 #include <cstddef>
