+++ src/OpenLoco/src/Ui/UiTypes.hpp
