+++ src/OpenLoco/src/Vehicles/Routing.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Limits.h"
+#include "Engine/Limits.h"
 #include <cstdint>
 
 namespace OpenLoco::Vehicles
